+++ cpp/core/src/zxing/pdf417/detector/LinesSampler.cpp
@@ -152,7 +152,7 @@ Ref<BitMatrix> LinesSampler::sample() {
   detectedCodeWords.resize(rowCount);
 
   // XXX
-  Ref<BitMatrix> grid(new BitMatrix(dimension_, detectedCodeWords.size()));
+  Ref<BitMatrix> grid(new BitMatrix(dimension_, (int) detectedCodeWords.size()));
   codewordsToBitMatrix(detectedCodeWords, grid);
 
   return grid;
